remove types separate file
